+++ src/app/(signin)/_components/signin-passkey.tsx
@@ -13,7 +13,7 @@ import {
   generatePasskeyAuthenticationOptionsAction,
   requestSessionAction,
   confirmSessionAction,
-} from "@/app/signin/actions";
+} from "@/app/(signin)/actions";
 import { toast } from "@/components/ui/use-toast";
 import * as simpleWebAuthn from "@simplewebauthn/browser";
 import { useSessionStore } from "@/state/session/state";
